Add a new function levenshtein_strcmp(), that compare two strings and
returns the levenshtein distance and g_strstr_lv() that searches in the string
haystack for the first occurrence of the string needle considering a maximum
levenshtein distance.

This allows approximate searches. =)
For example:
 * Search "Bjork" in "Enjoy - Björk - Post",
 * Search without writing the accents.
 * Search AC/DC AC-DC or AC DC regardless of how we write.
 * Large etc..

P.s: With this, searches are slower, and depending on the terms to be
searched, you can get unexpected results, but do not panic. Will be optional.