Fix: Some tracks are being displayed as unknown artist & album.
g_utf8_strlen (const gchar *p, gssize max);
The documentation clearly says.: max the maximum number of bytes to
examine. BYTES!! ! ! No utf8 characters! ! !.
And also says: If the max'th byte falls in the middle of a character, the last
(partial) character is not counted!.
Then, change max from 1 to 4 (The maximum number of bytes of one utf8
character).

P.s.: Many problems to save a little CPU and memory!.