Fixing duration formatting
Fixing issues with formatting of durations, e.g. seconds going up to
00:60 instead of switching to 1:00, or 1 hour and 22 seconds being
displayed as 1:022 instead of 1:00:22.

IMPLEMENTATION DETAILS

The third argument `dest_size` of

  ```
  gsize g_strlcat(gchar* dest, const gchar* src, gsize dest_size)

  ```

is the total size of of the buffer `dest`, not its remaining size or the
maximum size of `src` as seems to be the previous assumption.

In this context, the maximum length of the duration is 24 char (incl
null) and not 128 as allocated. I'm however relunctant to change either,
as I am not sure whether the buffer might be used elsewhere of if it is
some kind of optimization.