Replace bashisms in the configure scripts
Replace the use of `+=` operator with plain assignment to fix
compatibility with shells other than bash.  Without this script,
the build fails on systems where /bin/sh is not bash (and e.g. dash).

The original code leads to errors such as:

    ./configure: 14711: PRAGHA_CFLAGS+= ... : not found

since the assignment is wrongly taken to be a command to execute.
As a result, flags are not being appended and the build eventually
fails with e.g.:

    pragha-cmdline.c:33:10: fatal error: libxfce4ui/libxfce4ui.h: No such file or directory

Original bug report: https://bugs.gentoo.org/889618